+++ packages/amiddy/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ## v-next
 * :house: moved project source to monorepo (from https://github.com/darkyndy/amiddy)
+* :rocket: Ability to set custom response headers for every request
 
 
 ## v1.3.3